Add mechanism to set headers in Stub.connect/3
This patch provides a generic way to set headers that will be attached
to each request from this `channel` by re-using
append_custom_metadata/2.

The use case is to provide a location for setting authorization headers
that does not require manually adding it to each Stub fn call.

With this change, headers can be set as:

```
channel = Stub.connect(host, port, headers: [{"authorization", "Bearer TOKEN"}])
```

On the receiving server side they can be accessed using existing getter
function. For the authorization example, this can be done in an
Interceptor:

```elixir
defmodule Authenticator do
  def init(_), do: []
  def call(req, stream, next, opts) do
    with headers when is_map(headers) <- GRPC.Stream.get_headers(stream),
         token when not is_nil(token) <- Map.get(headers, "authorization") do
         .... # use token
         .... # if authenticated
         next.(req, stream)
    end
  end
end
```

Adds two tests, one to verify on client side that headers can be set and
accessible, a second test is for integration to verify the roundtrip of
a header from client to server, then encode the header into a GRPC
message response for test verification.