Allow RPC handlers to return errors
Having to raise an exception to signal an error is slower (because
exceptions are, well, exceptional), and it also interferes with
exception logging interceptors.  This change allows RPC handler
methods to return {:error, ...} and have the expected thing happen --
which happens to be exactly how interceptors work, as a bonus.