Access mode detection (transcode vs grpcweb vs grpc) (#412)
* Detect preflight requests

These typically come from grpcweb clients doing OPTIONS checks for CORS

* Detect client types: grpc, grpcweb, web

This replaces the `http_transcode` member of the `Server.Stream` struct
as transcoding is correlated with being a web client (*not* a grpcweb
client, however!)

Knowing the type of client also allows for detection of when to send
different types of headers, for e.g. CORS

* Add a CORS header injection Interceptor

It an option `:allow` option to define which origins are allowed.

This may be a static string or a function which will be passed the
current `req` and `stream` structs to choose what to allow.

* Reintroduce the boolean http_transcode member of the Stream struct

NOTE: will merge this with commit
c39c83498cfe23fea54d384c06ae7510a840b276 post-review/discussion.
I am leaving it as a separate commit for now for ease of continued
discussion on the current PR.

* web can be detected in extract_subtype

* change access_type to access_mode, :web to :http_transcoding

this more accurately and directly maps to the grpc terminology, and is
probably clearer.

* default opts to an empty list, whitespace fix

* add tests for the CORS header interceptor

* Assign to a temporary variable and assign that to Stream.http_transcode

* provide access to request headers in the Stream object

* only set cors when the sec-fetch-mode header indicates it

* another test: if there is NO sec-fetch-mode header

* Make access-control-allow-headers configurable, and optional

Only send access-control-allow-headers when the client requests it with
access-control-request-headers and allow the same configurability as
access-control-allow-origin provides

* align variable names with http header names

* Add a CORS example to the README

* format

* Verbage.

Co-authored-by: Paulo Valente <16843419+polvalente@users.noreply.github.com>

* remove default for allow_origin in CORS interceptor

* raise on failed allow_origin type, and test for that

* allow_origin, not allow_header

* test: add missing compile-time test

* chore: format readme

* chore: fix Note bold

* docs: cleanup readme linking

* docs: proper module linking in exdoc

* fix: do not use Keyword.validate due to Elixir 1.12 support

---------

Co-authored-by: Adriano Santos <solid.sistemas@gmail.com>
Co-authored-by: Paulo Valente <16843419+polvalente@users.noreply.github.com>