Give the status field a default value of unknown
This avoids a foot-gun where the app dev does not know which status to use
and only sets the message (and/or potentially the details). Without a
default value, the status field becomes `nil` which blows up when
serializing which is not an easy bug to trace back to the right place in
the application.