the next action taken can depend on the request being grpcweb
when the request isn't a grpcweb request, then it's business as usual.
otherwise, the grpcweb trailers must be sent first, as they may cause a
body to be sent.

only after checking for grpcweb can the regular trailers be sent once
the state of the req is confirmed, namely whether or not a reply has
been started already or if a full reply must be initiated.